jsonurl_py.py: Rename to avoid conflict with https://pypi.org/project/jsonurl
